add alert checker classes for monitoring alerts and refactor imports for improved structure
